Don't recommend a single useMappedState call
Per issue #12, there is not a strong performance-based reason to caution against multiple uses of useMappedState in a single component, particularly when store fetches are abstracted into a custom hook.

Update the readme to reflect that multiple calls to useMappedState will not rerender the component multiple times, but may have an impact due to the number of subscriptions in Redux.