Allowing use of memoization array.
As explained in #30 

Hooks usually accept a second argument of values to memoize over. In that respect, the `useMappedState` is a little weird in that it accepts the mapper function to be pre-memoized or defined as a constant outside of the component.

Accepting an argument of values to memoize based on removes the need to do that.

By defaulting to reference to the mapper function, we get existing semantics when no second argument is provided and the change is backwards compatible.