subscribe parent first
I'm not sure about this. This was just a quick and dirty thing that popped into my head. I will try to test if this works sometime.

But I think this should hopefully ensure that at least on the first render, a parent will subscribe to changes before the child.
Things get messy if the store needs to resubscribe for some reason.